common: Split common file
